Reorganized directory and made configureTwitterBot return a TwitterBot instance instead of the constructor itself for usability.
Moved validate and custom-validators into lib and moved schemas folder to root.

Also refactored validate so that it could inherit options passed into configureTwitterBot.

configureTwitterBot now returns an instance of TwitterBot with no rules, but otherwise inheriting options passed into configureTwitterBot. This was done to eliminate the extra step of having to call the constructor function after passing in the config info, which seemed like an unnecessary complication.