Refactored validator and rewrote tests
Removed ability to inject dependencies because they offer no utility for testing and could allow a user to cause errors. Also switched from a factory function to exporting a single instance so that the singleton would be shared everywhere it was required. Now it can be used by other modules without having to inject it into everything else.